requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mysqlclient==2.2.4
python-dotenv==1.0.1
SQLAlchemy==2.0.28
typing_extensions==4.10.0
Werkzeug==3.0.1
